amazon wishlist <-> (monroe county library system, rit library) scraper.
i use my amazon wishlist as a repository for things i might want to read in
future, recommendations from friends, things i might want to check out of
the library, etc. it's nice to have something spit out a list of library
availability and amazon.com new/used prices so i can decide how to obtain an
item ("hey, i'm going to be near the library today, i'll pick $X up since
it's available", "nobody around here has $X but it's available used for
cheap, i'll just buy it")

i've been using this for a while, and i spent a couple hours putting a
little polish on it today. it's still not the prettiest/most robust/best
documented thing ever, but i wanted a place to revision control it and
figured someone else might find it interesting or useful.